refactor: split daemon foundation modules
